Travis: Refactor out ds3_share_open_twice
Large methods need to be in separate modules.
